requirements.txt
python-dotenv~=0.21.0
beautifulsoup4~=4.11.1
schedule~=1.1.0